+++ employeeservice/src/main/java/com/example/employeeservice/model/Department.java
@@ -1,22 +1,22 @@
 package com.example.employeeservice.model;
 
 public class Department {
-    private Integer id;
+    private Long id;
     private String name;
     
     public Department() {
     }
-    public Department(Integer id, String name) {
+    public Department(Long id, String name) {
         this.id = id;
         this.name = name;
     }
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
     public String getName() {
         return name;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public void setName(String name) {
