+++ employeeservice/src/main/java/com/example/employeeservice/model/Employee.java
@@ -1,7 +1,7 @@
 package com.example.employeeservice.model;
 
 public class Employee {
-    private Integer id;
+    private Long id;
     private String firstName;
     private String lastName;
     private String email;
@@ -10,7 +10,7 @@ public class Employee {
     
     public Employee() {
     }
-    public Employee(Integer id, String firstName, String lastName, String email, String phone, Department department) {
+    public Employee(Long id, String firstName, String lastName, String email, String phone, Department department) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -18,7 +18,7 @@ public class Employee {
         this.phone = phone;
         this.department = department;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public void setFirstName(String firstName) {
@@ -36,7 +36,7 @@ public class Employee {
     public void setDepartment(Department department) {
         this.department = department;
     }
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
     public String getFirstName() {
